+++ tests/testthat/test-linear_reg_quantreg.R
@@ -1,14 +1,7 @@
 test_that('linear quantile regression via quantreg - single quantile', {
   skip_if_not_installed("quantreg")
 
-  data("Sacramento")
-
-  Sacramento_small <-
-    Sacramento %>%
-    dplyr::select(price, beds, baths, sqft, latitude, longitude)
-
-  sac_train <- Sacramento_small[-(1:5), ]
-  sac_test  <- Sacramento_small[  1:5 , ]
+  # data in `helper-objects.R`
 
   one_quant <-
     linear_reg() %>%
@@ -24,9 +17,18 @@ test_that('linear quantile regression via quantreg - single quantile', {
   expect_true(nrow(one_quant_pred) == nrow(sac_test))
   expect_named(one_quant_pred, ".pred_quantile")
   expect_true(is.list(one_quant_pred[[1]]))
-  expect_s3_class(one_quant_pred$.pred_quantile[[1]], c("tbl_df", "tbl", "data.frame"))
-  expect_named(one_quant_pred$.pred_quantile[[1]], c(".pred_quantile", ".quantile_level"))
-  expect_true(nrow(one_quant_pred$.pred_quantile[[1]]) == 1L)
+  expect_s3_class(
+    one_quant_pred$.pred_quantile[1],
+    c("quantile_pred", "vctrs_vctr", "list")
+  )
+  expect_identical(class(one_quant_pred$.pred_quantile[[1]]), "numeric")
+  expect_true(length(one_quant_pred$.pred_quantile[[1]]) == 1L)
+  expect_identical(attr(one_quant_pred$.pred_quantile, "quantile_levels"), .5)
+
+  one_quant_df <- as_tibble(one_quant_pred$.pred_quantile)
+  expect_s3_class(one_quant_df, c("tbl_df", "tbl", "data.frame"))
+  expect_named(one_quant_df, c(".pred_quantile", ".quantile_levels", ".row"))
+  expect_true(nrow(one_quant_df) == nrow(sac_test) * 1)
 
   ###
 
@@ -34,22 +36,24 @@ test_that('linear quantile regression via quantreg - single quantile', {
   expect_true(nrow(one_quant_one_row) == 1L)
   expect_named(one_quant_one_row, ".pred_quantile")
   expect_true(is.list(one_quant_one_row[[1]]))
-  expect_s3_class(one_quant_one_row$.pred_quantile[[1]], c("tbl_df", "tbl", "data.frame"))
-  expect_named(one_quant_one_row$.pred_quantile[[1]], c(".pred_quantile", ".quantile_level"))
-  expect_true(nrow(one_quant_one_row$.pred_quantile[[1]]) == 1L)
+  expect_s3_class(
+    one_quant_one_row$.pred_quantile[1],
+    c("quantile_pred", "vctrs_vctr", "list")
+  )
+  expect_identical(class(one_quant_one_row$.pred_quantile[[1]]), "numeric")
+  expect_true(length(one_quant_one_row$.pred_quantile[[1]]) == 1L)
+  expect_identical(attr(one_quant_pred$.pred_quantile, "quantile_levels"), .5)
+
+  one_quant_one_row_df <- as_tibble(one_quant_one_row$.pred_quantile)
+  expect_s3_class(one_quant_one_row_df, c("tbl_df", "tbl", "data.frame"))
+  expect_named(one_quant_one_row_df, c(".pred_quantile", ".quantile_levels", ".row"))
+  expect_true(nrow(one_quant_one_row_df) == nrow(sac_test[1,]) * 1)
 })
 
 test_that('linear quantile regression via quantreg - multiple quantiles', {
   skip_if_not_installed("quantreg")
 
-  data("Sacramento")
-
-  Sacramento_small <-
-    Sacramento %>%
-    dplyr::select(price, beds, baths, sqft, latitude, longitude)
-
-  sac_train <- Sacramento_small[-(1:5), ]
-  sac_test  <- Sacramento_small[  1:5 , ]
+  # data in `helper-objects.R`
 
   ten_quant <-
     linear_reg() %>%
@@ -65,9 +69,18 @@ test_that('linear quantile regression via quantreg - multiple quantiles', {
   expect_true(nrow(ten_quant_pred) == nrow(sac_test))
   expect_named(ten_quant_pred, ".pred_quantile")
   expect_true(is.list(ten_quant_pred[[1]]))
-  expect_s3_class(ten_quant_pred$.pred_quantile[[1]], c("tbl_df", "tbl", "data.frame"))
-  expect_named(ten_quant_pred$.pred_quantile[[1]], c(".pred_quantile", ".quantile_level"))
-  expect_true(nrow(ten_quant_pred$.pred_quantile[[1]]) == 10L)
+  expect_s3_class(
+    ten_quant_pred$.pred_quantile[1],
+    c("quantile_pred", "vctrs_vctr", "list")
+  )
+  expect_identical(class(ten_quant_pred$.pred_quantile[[1]]), "numeric")
+  expect_true(length(ten_quant_pred$.pred_quantile[[1]]) == 10L)
+  expect_identical(attr(ten_quant_pred$.pred_quantile, "quantile_levels"), (0:9)/9)
+
+  ten_quant_df <- as_tibble(ten_quant_pred$.pred_quantile)
+  expect_s3_class(ten_quant_df, c("tbl_df", "tbl", "data.frame"))
+  expect_named(ten_quant_df, c(".pred_quantile", ".quantile_levels", ".row"))
+  expect_true(nrow(ten_quant_df) == nrow(sac_test) * 10)
 
   ###
 
@@ -75,9 +88,21 @@ test_that('linear quantile regression via quantreg - multiple quantiles', {
   expect_true(nrow(ten_quant_one_row) == 1L)
   expect_named(ten_quant_one_row, ".pred_quantile")
   expect_true(is.list(ten_quant_one_row[[1]]))
-  expect_s3_class(ten_quant_one_row$.pred_quantile[[1]], c("tbl_df", "tbl", "data.frame"))
-  expect_named(ten_quant_one_row$.pred_quantile[[1]], c(".pred_quantile", ".quantile_level"))
-  expect_true(nrow(ten_quant_one_row$.pred_quantile[[1]]) == 10L)
+  expect_s3_class(
+    ten_quant_one_row$.pred_quantile[1],
+    c("quantile_pred", "vctrs_vctr", "list")
+  )
+  expect_identical(class(ten_quant_one_row$.pred_quantile[[1]]), "numeric")
+  expect_true(length(ten_quant_one_row$.pred_quantile[[1]]) == 10L)
+  expect_identical(
+    attr(ten_quant_one_row$.pred_quantile, "quantile_levels"),
+    (0:9)/9
+  )
+
+  ten_quant_one_row_df <- as_tibble(ten_quant_one_row$.pred_quantile)
+  expect_s3_class(ten_quant_one_row_df, c("tbl_df", "tbl", "data.frame"))
+  expect_named(ten_quant_one_row_df, c(".pred_quantile", ".quantile_levels", ".row"))
+  expect_true(nrow(ten_quant_one_row_df) == nrow(sac_test[1,]) * 10)
 })